Merilife/MediBot/requirements.txt
# Django==5.0.1
# django-cors-headers==4.3.1
# django-rest-framework==0.1.0
# django-environ==0.11.2
# python-dotenv==1.0.0
# django-allauth==0.60.0
# django-guardian==2.4.0
# pymongo==4.6.1
# sentence-transformers==2.2.2
# torch==2.1.2
# transformers==4.36.2
# python-jose==3.3.0
# passlib==1.7.4
# python-multipart==0.0.6
# PyJWT==2.8.0
# Pillow==10.2.0
# pandas==2.1.4
# numpy==1.26.3
# matplotlib==3.8.2
# seaborn==0.13.1
# python-docx==1.1.0
# reportlab==4.0.9
# flake8==7.0.0
# black==23.12.1
# pytest==7.4.4
# pytest-django==4.7.0
# pytest-cov==4.1.0

absl-py==2.1.0
aider-chat==0.60.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
bitcoinlib==0.6.15
bleach==6.2.0
blinker==1.8.2
blis==0.7.11
boto3==1.34.127
botocore==1.34.127
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.20.0
cmsisdsp==1.9.9
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.3.0
cookiecutter==2.6.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
debugpy==1.8.5
decorator==4.4.2
defang==0.5.3
diff-match-patch==20230430
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.1
django-environ==0.12.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.15
duckdb==1.1.3
easyocr==1.7.2
ecdsa==0.19.0
elevenlabs==1.3.0
emoji==1.7.0
et_xmlfile==2.0.0
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-MySQLdb==1.0.1
flatbuffers==24.12.23
folium==0.19.5
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.1
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.0.3
grep-ast==0.3.3
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.1
gunicorn==23.0.0
gurobipy==11.0.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.25.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.1
importlib_resources==6.4.5
instructor==0.6.8
intel-openmp==2021.4.0
ipaddress==1.0.23
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.17
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langcodes==3.4.1
langsmith==0.3.5
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
litellm==1.47.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.10
mutagen==1.47.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
openai==0.28.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatasets==0.1.22
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.15
packaging==24.1
pandas==2.2.2
pandasql==0.7.3
parso==0.8.4
pathos==0.3.3
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
playwright==1.44.0
pox==0.3.5
ppft==1.7.6.9
praw==7.7.1
prawcore==2.3.0
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
PuLP==2.9.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==11.1.0
PyExecJS==1.5.1
pyflakes==3.2.0
PyGetWindow==0.0.9
Pygments==2.18.0
pymongo==4.11
PyMsgBox==1.0.9
pynndescent==0.5.13
pypandoc==1.13
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-bitcoinlib==0.12.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.2
referencing==0.35.1
regex==2024.9.11
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.5.2
scikit-posthocs==0.11.2
scipy==1.13.1
seaborn==0.13.2
setuptools==75.8.0
shapely==2.0.7
shell_gpt==1.4.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tbb==2021.13.1
tenacity==8.5.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2025.3.13
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.41.2
translators==5.9.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.9.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.6
Unidecode==1.3.8
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2025.1.0
yarl==1.11.1
yt-dlp==2024.5.27
zipp==3.20.2
zstandard==0.23.0